Adjusted per Miro's comments.
